requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
brotli==1.0.9
    # via yt-dlp
certifi==2022.6.15
    # via
    #   requests
    #   yt-dlp
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
mutagen==1.45.1
    # via yt-dlp
pycryptodomex==3.14.1
    # via yt-dlp
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
urllib3==1.26.9
    # via requests
websockets==10.3
    # via yt-dlp
yt-dlp==2022.5.18
    # via -r requirements.in
